Removing `output-map` from verbose-console-reporter
This is the first in a series of refactors to verbose-console-reporter. There will also be a change where some additional util functions (similar to `renderLintingSummaryResult`) exposed to offer an additional layer of options between using ui library and building result per task directly, and just using the current default behavior.

This change simply removes `output-map` and all task specific code from verbose-console-reporter, and moves it to the plugins directly. Our internal plugins should register custom reporters the same way we would expect an external plugin to. There are no functionality changes in this PR, just moving existing custom cases (where necessary) to plugins
