+++ checkup.code-workspace
@@ -13,7 +13,8 @@
   },
   "settings": {
     "files.exclude": {
-      "**/node_modules/**": true
+      "**/node_modules/**": true,
+      "**/lib/**": true
     },
     "typescript.tsdk": "node_modules/typescript/lib",
     "editor.formatOnSave": true
