+++ packages/checkup-plugin-ember/src/reporters/ember-template-lint-summary-reporter.ts
@@ -1,0 +1,6 @@
+import { renderLintingSummaryResult } from '@checkup/core';
+import { Result } from 'sarif';
+
+export function report(taskResults: Result[]) {
+  renderLintingSummaryResult(taskResults);
+}
