+++ packages/checkup-plugin-javascript/package.json
@@ -40,7 +40,8 @@
     ],
     "hooks": {
       "register-info-tasks": "./lib/hooks/register-info-tasks",
-      "register-actions": "./lib/hooks/register-actions"
+      "register-actions": "./lib/hooks/register-actions",
+      "register-task-reporter": "./lib/hooks/register-task-reporters"
     }
   },
   "publishConfig": {
