+++ packages/cli/src/reporters/verbose-console-reporter.ts
@@ -7,107 +7,12 @@ import {
   NO_RESULTS_FOUND,
   sumOccurrences,
   reduceResults,
+  renderEmptyResult,
 } from '@checkup/core';
 import * as cleanStack from 'clean-stack';
 import { startCase } from 'lodash';
 import { Invocation, Log, Notification, Result, Run } from 'sarif';
 import { renderActions, renderCLIInfo, renderInfo, renderLinesOfCode } from './reporter-utils';
-
-let outputMap: { [taskName: string]: (taskResults: Result[]) => void } = {
-  'ember-dependencies': function (taskResults: Result[]) {
-    if (!taskResults.some((result: Result) => result.properties?.data !== undefined)) {
-      return;
-    }
-
-    ui.section(taskResults[0].properties?.taskDisplayName, () => {
-      taskResults.forEach((result: Result) => {
-        if (result.message.text === NO_RESULTS_FOUND) {
-          renderEmptyResult(result);
-        } else {
-          ui.value({ title: result.message.text, count: result.properties?.data.length });
-        }
-      });
-    });
-  },
-  'ember-test-types': function (taskResults: Result[]) {
-    let groupedTaskResults = groupDataByField(taskResults, 'message.text');
-
-    ui.section(taskResults[0].properties?.taskDisplayName, () => {
-      groupedTaskResults.forEach((resultGroup: Result[]) => {
-        let groupedTaskResultsByMethod = reduceResults(
-          groupDataByField(resultGroup, 'properties.method')
-        );
-        ui.subHeader(groupedTaskResultsByMethod[0].message.text);
-        ui.valuesList(
-          groupedTaskResultsByMethod.map((result) => {
-            if (result.message.text === NO_RESULTS_FOUND) {
-              renderEmptyResult(result);
-            } else {
-              return { title: result.properties?.method, count: result.occurrenceCount };
-            }
-          })
-        );
-        ui.blankLine();
-      });
-
-      ui.subHeader('tests by type');
-      ui.sectionedBar(
-        groupedTaskResults.map((results: Result[]) => {
-          return {
-            title: results[0].message.text,
-            count: sumOccurrences(results),
-          };
-        }),
-        sumOccurrences(taskResults),
-        'tests'
-      );
-    });
-  },
-  'ember-octane-migration-status': function (taskResults: Result[]) {
-    ui.section(taskResults[0].properties?.taskDisplayName, () => {
-      ui.log(`${ui.emphasize('Octane Violations')}: ${sumOccurrences(taskResults)}`);
-      ui.blankLine();
-
-      let groupedTaskResults = groupDataByField(taskResults, 'properties.resultGroup');
-
-      groupedTaskResults.forEach((resultGroup: Result[]) => {
-        let groupedTaskResultsByLintRuleId = reduceResults(
-          groupDataByField(resultGroup, 'properties.lintRuleId')
-        );
-
-        ui.subHeader(groupedTaskResultsByLintRuleId[0].properties?.resultGroup);
-        ui.valuesList(
-          groupedTaskResultsByLintRuleId.map((result) => {
-            if (result.message.text === NO_RESULTS_FOUND) {
-              renderEmptyResult(result);
-            } else {
-              return { title: result.properties?.lintRuleId, count: result.occurrenceCount };
-            }
-          }),
-          'violations'
-        );
-        ui.blankLine();
-      });
-    });
-  },
-  'ember-template-lint-summary': renderLintingSummaryResult,
-  'eslint-summary': renderLintingSummaryResult,
-  'outdated-dependencies': function (taskResults: Result[]) {
-    ui.section(taskResults[0].properties?.taskDisplayName, () => {
-      ui.sectionedBar(
-        taskResults.map((result: Result) => {
-          if (result.message.text === NO_RESULTS_FOUND) {
-            renderEmptyResult(result);
-          } else {
-            return { title: result.message.text, count: result.occurrenceCount };
-          }
-        }),
-        sumOccurrences(taskResults),
-        'dependencies'
-      );
-    });
-  },
-};
 
 export function report(result: Log) {
   let metaData = result.properties as CheckupMetadata;
@@ -156,7 +61,7 @@ function renderTaskResults(pluginTaskResults: Result[] | undefined): void {
 function getTaskReporter(taskResult: Result[]) {
   let taskName = taskResult[0].ruleId as string;
   let registeredTaskReporters = getRegisteredTaskReporters();
-  let reporter = outputMap[taskName] || registeredTaskReporters.get(taskName) || getReportComponent;
+  let reporter = registeredTaskReporters.get(taskName) || getReportComponent;
 
   if (typeof reporter === 'undefined') {
     throw new CheckupError(
@@ -166,13 +71,6 @@ function getTaskReporter(taskResult: Result[]) {
   }
 
   return reporter;
-}
-
-export function renderEmptyResult(taskResult: Result) {
-  ui.value({
-    title: taskResult.properties?.consoleMessage || taskResult.properties?.taskDisplayName,
-    count: 0,
-  });
 }
 
 function getReportComponent(taskResults: Result[]) {
@@ -234,30 +132,4 @@ function renderTimings(timings: Record<string, number>) {
     }
   );
   ui.blankLine();
-}
-
-function renderLintingSummaryResult(taskResults: Result[]) {
-  let groupedTaskResultsByType = groupDataByField(taskResults, 'properties.type');
-
-  ui.section(taskResults[0].properties?.taskDisplayName, () => {
-    groupedTaskResultsByType.forEach((resultGroup: Result[]) => {
-      let groupedTaskResultsByLintRule = reduceResults(
-        groupDataByField(resultGroup, 'properties.lintRuleId')
-      ).sort((a, b) => (b.occurrenceCount || 0) - (a.occurrenceCount || 0));
-      let totalCount = sumOccurrences(groupedTaskResultsByLintRule);
-      if (totalCount) {
-        ui.subHeader(`${groupedTaskResultsByLintRule[0].properties?.type}s: (${totalCount})`);
-        ui.valuesList(
-          groupedTaskResultsByLintRule.map((result) => {
-            if (result.message.text === NO_RESULTS_FOUND) {
-              renderEmptyResult(result);
-            } else {
-              return { title: result.properties?.lintRuleId, count: result?.occurrenceCount };
-            }
-          })
-        );
-        ui.blankLine();
-      }
-    });
-  });
 }
