+++ packages/core/src/index.ts
@@ -32,6 +32,7 @@ export { ui } from './utils/ui';
 export { getFilePaths, getFilePathsAsync } from './utils/get-paths';
 export { FilePathArray } from './utils/file-path-array';
 export { sumOccurrences, reduceResults } from './utils/sarif-utils';
+export { renderEmptyResult, renderLintingSummaryResult } from './utils/verbose-reporter-utils';
 
 export { byRuleId, byRuleIds, bySeverity } from './data/filters';
 export { toPercent, groupDataByField } from './data/formatters';
